Initial commit. Includes current version of the Header APP with all dependences.
The Header APP is SPA. It has a header with a menu and an app launcher button and a body with a form rendered by Form.IO js component. JSON representation of the form is stored in form.json.js. The APP is able to sign users with a Microsoft accounts in and to fetch their username, email address and photo and fetch and update their mailbox language data. Language setting is applied to the APP. Users can also apply different theme and see focused form component and global form help. Form components can be pre-populated with some user's values.