Fix for an intermittent issue with loading the header application.
Bug description: Our definition files are loaded dynamically. We first find their paths and then include them to the app by adding a new script element to the head part of the index.html. So they are loaded asychronously and it is possible that they are not loaded yet when window load event occurs.

Bug fix: We can prevent jquery document ready event to be fired until we load all of our definitions files by adding $.holdReady( true ) for each definition file to be executed before the file has been loaded and adding $.holdReady( false ) at the end of each definition file. Then we can add listener to document ready and run the app setup script if the page has already been loaded or add the setup script as a listener for window load event.