Fix for an intermittent issue with loading the header application. $.holdReady statements are removed from our *.json.js def files.
Bug description: Our definition files are loaded dynamically. They need to be loaded one by one since we have dependances between them or otherwise wrong definitions can be loaded.

Bug fix: We define callbacks when script elements are loaded or when their load fails. Then when we have all scripts loaded we check if document is already loaded. If so we start to setup the APP. Otherwise we add a document onload listener which starts the app setup.