Available themes can be specified in themes.json.js definition file.
In header.json.js the "themes" parameter defines the path to the themes.json.js file where available themes can be defined on the similar way as in theme.js file. If the "themes" parameter in header.json.js hasn't been defined or has an empty value, the hardcoded themes are used.